Allow pattern.standard_FPR to take len-1 patterns
When `reference_patterns` or `estimated_patterns` were both of length 1,
then np.diff(P - Q, axis=0) would return an array of shape `(0, 2)`
which would cause a `ValueError` from `np.max`.  This test was basically
just that the change in semitones in the reference and estimated
patterns was the same.  When there's only one note in the pattern, the
change in semitones is the same, so we just add the one-note case as a
special case.