Add key detection evaluation
This adds the single key detection eval function used by MIREX (a
heuristic score which weights certain key results as better/worse than
others) and an io function for loading in data in the MIREX format (a
single tab-separated line denoting the key scale degree and mode).  Its
output was verified against a Python port of the IMIRSEL code.