Loosen separation test tolerance (#302)
For some tests, the difference between the scores produced by Travis'
build and my own build is as large as 3.89588121e-04. In practice, we
shouldn't care much about differences larger than 10e-3 for separation
tasks since it would be very unusual for anyone to pay attention to
differences this small when comparing source separation algorithms.